# sdks/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMSafeSpaces API
  description: |
    LLMSafeSpaces is a Kubernetes-first platform for running AI agents securely in isolated sandboxes.
    Every sandbox runs `opencode serve` as a persistent HTTP server with a PVC-backed persistent workspace.
  version: "0.7.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /secrets:
    post:
      tags: [secrets]
      summary: Create a secret
      operationId: createSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
      responses:
        "201":
          description: Secret created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [secrets]
      summary: List secrets
      operationId: listSecrets
      responses:
        "200":
          description: Secret list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /secrets/{id}:
    get:
      tags: [secrets]
      summary: Get a secret (value not included)
      operationId: getSecret
      parameters:
        - $ref: "#/components/parameters/SecretId"
      responses:
        "200":
          description: Secret metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [secrets]
      summary: Update a secret value
      operationId: updateSecret
      parameters:
        - $ref: "#/components/parameters/SecretId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecretRequest"
      responses:
        "200":
          description: Secret updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [secrets]
      summary: Delete a secret
      operationId: deleteSecret
      parameters:
        - $ref: "#/components/parameters/SecretId"
      responses:
        "204":
          description: Secret deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{id}/env:
    put:
      tags: [secrets]
      summary: Set workspace environment variables
      operationId: setWorkspaceEnv
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Environment set
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [secrets]
      summary: Get workspace environment variables
      operationId: getWorkspaceEnv
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Environment variables
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{id}/env/{name}:
    delete:
      tags: [secrets]
      summary: Delete a workspace environment variable
      operationId: deleteWorkspaceEnv
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Environment variable deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  # --- Settings ---
...
components:
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
...
    SecretId:
      name: id
      in: path
      required: true
      description: Secret ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or /auth/register
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "lsp_ prefixed API key in Bearer format (e.g. Bearer lsp_abc123...)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateSecretRequest:
      type: object
      required: [name, type, value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9._-]+$'
          description: Lowercase alphanumeric, dots, underscores, hyphens only. Must not start with dot or hyphen.
        type:
          type: string
          enum: [llm-provider, ssh-key, git-credential, secret-file, env-secret]
        value:
          type: string
          description: Plaintext value (encrypted before storage)
        metadata:
          type: object
          description: Arbitrary metadata JSON
    UpdateSecretRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
        metadata:
          type: object
    SecretResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [llm-provider, ssh-key, git-credential, secret-file, env-secret]
        metadata:
          type: object
        globalDefault:
          type: boolean
          description: Whether the secret applies to all workspaces by default; always present (no omitempty).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...